refactor(bench): extract shared utilities to bench_utils module (#20)
Move duplicated code from locomo/ and longmemeval/ into a shared
benches/bench_utils/ module:
- metrics.rs: PeakRss + current_rss_kb (macOS, Linux, fallback)
- formatting.rs: pct, grade, truncate (with unit tests)
- openai_types.rs: OpenAI chat request/response structs (6 types)

Both bench binaries import via #[path] attribute. Domain-specific
functions (substring_match, record_result, category lists) remain
in their respective benches.
